Move SetOperationNodes to SPI
