+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/ExceptNode.java
@@ -11,12 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.sql.planner.plan;
+package com.facebook.presto.spi.plan;
 
-import com.facebook.presto.spi.plan.PlanNode;
-import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.annotation.concurrent.Immutable;
@@ -25,11 +22,10 @@ import java.util.List;
 import java.util.Map;
 
 @Immutable
-public class UnionNode
+public class ExceptNode
         extends SetOperationNode
 {
-    @JsonCreator
-    public UnionNode(
+    public ExceptNode(
             @JsonProperty("id") PlanNodeId id,
             @JsonProperty("sources") List<PlanNode> sources,
             @JsonProperty("outputToInputs") Map<VariableReferenceExpression, List<VariableReferenceExpression>> outputToInputs)
@@ -38,14 +34,14 @@ public class UnionNode
     }
 
     @Override
-    public <R, C> R accept(InternalPlanVisitor<R, C> visitor, C context)
+    public <R, C> R accept(PlanVisitor<R, C> visitor, C context)
     {
-        return visitor.visitUnion(this, context);
+        return visitor.visitExcept(this, context);
     }
 
     @Override
     public PlanNode replaceChildren(List<PlanNode> newChildren)
     {
-        return new UnionNode(getId(), newChildren, getVariableMapping());
+        return new ExceptNode(getId(), newChildren, getVariableMapping());
     }
 }
