+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/IntersectNode.java
@@ -11,11 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.sql.planner.plan;
+package com.facebook.presto.spi.plan;
 
-import com.facebook.presto.spi.plan.PlanNode;
-import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.annotation.concurrent.Immutable;
@@ -24,10 +23,11 @@ import java.util.List;
 import java.util.Map;
 
 @Immutable
-public class ExceptNode
+public class IntersectNode
         extends SetOperationNode
 {
-    public ExceptNode(
+    @JsonCreator
+    public IntersectNode(
             @JsonProperty("id") PlanNodeId id,
             @JsonProperty("sources") List<PlanNode> sources,
             @JsonProperty("outputToInputs") Map<VariableReferenceExpression, List<VariableReferenceExpression>> outputToInputs)
@@ -36,14 +36,14 @@ public class ExceptNode
     }
 
     @Override
-    public <R, C> R accept(InternalPlanVisitor<R, C> visitor, C context)
+    public <R, C> R accept(PlanVisitor<R, C> visitor, C context)
     {
-        return visitor.visitExcept(this, context);
+        return visitor.visitIntersect(this, context);
     }
 
     @Override
     public PlanNode replaceChildren(List<PlanNode> newChildren)
     {
-        return new ExceptNode(getId(), newChildren, getVariableMapping());
+        return new IntersectNode(getId(), newChildren, getVariableMapping());
     }
 }
