+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/UnionNode.java
@@ -11,10 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.sql.planner.plan;
+package com.facebook.presto.spi.plan;
 
-import com.facebook.presto.spi.plan.PlanNode;
-import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -25,11 +23,11 @@ import java.util.List;
 import java.util.Map;
 
 @Immutable
-public class IntersectNode
+public class UnionNode
         extends SetOperationNode
 {
     @JsonCreator
-    public IntersectNode(
+    public UnionNode(
             @JsonProperty("id") PlanNodeId id,
             @JsonProperty("sources") List<PlanNode> sources,
             @JsonProperty("outputToInputs") Map<VariableReferenceExpression, List<VariableReferenceExpression>> outputToInputs)
@@ -38,14 +36,14 @@ public class IntersectNode
     }
 
     @Override
-    public <R, C> R accept(InternalPlanVisitor<R, C> visitor, C context)
+    public <R, C> R accept(PlanVisitor<R, C> visitor, C context)
     {
-        return visitor.visitIntersect(this, context);
+        return visitor.visitUnion(this, context);
     }
 
     @Override
     public PlanNode replaceChildren(List<PlanNode> newChildren)
     {
-        return new IntersectNode(getId(), newChildren, getVariableMapping());
+        return new UnionNode(getId(), newChildren, getVariableMapping());
     }
 }
